refactor: refactor elements to material design and start adaptive
