--- src/components/footer/footer.css
@@ -1,19 +1,0 @@
-.footer {
-    display: flex;
-    align-items: center;
-    height: 50px;
-    padding: 0 20px;
-    background-color: rgb(0, 0, 0);
-    font-size: 20px;
-}
-
-.footer_github-list {
-    display: flex;
-    gap: 10px;
-}
-  
-.github-list_link {
-    color: #ffffff;
-    text-decoration: none;
-    font-size: 0.9rem;
-}
